// coverage-app/src/types.ts
export interface CoverageItem {
  date?: string;
  duration_all?: number;
  volume_all?: number;
  session_count_all?: number;
  unique_imsi_count_all?: number;
  volte_drop_count?: number;
  volte_failure_count?: number;
  volte_normal_count?: number;
  dl_throuput_max?: number;
  unique_imsi_count_with_dl_throuput_max_lt1?: number;
  unique_imsi_count_with_dl_throuput_max_gte1_lt3?: number;
  unique_imsi_count_with_dl_throuput_max_gte3_lt5?: number;
  unique_imsi_count_with_dl_throuput_max_gte5_lt10?: number;
  unique_imsi_count_with_dl_throuput_max_gte10_lt20?: number;
  unique_imsi_count_with_dl_throuput_max_gte20_lt50?: number;
  unique_imsi_count_with_dl_throuput_max_gte50?: number;
  unique_imsi_count_oos?: number;
  unique_imsi_count_oos_with_session_count_1?: number;
  unique_imsi_count_oos_with_session_count_gt1_lte5?: number;
  unique_imsi_count_oos_with_session_count_gt5_lte10?: number | null;
  unique_imsi_count_oos_with_session_count_gt10_lte20?: number | null;
  unique_imsi_count_oos_with_session_count_gt20_lte50?: number | null;
  unique_imsi_count_oos_with_session_count_gt50_lte100?: number | null;
  unique_imsi_count_oos_with_session_count_gt100?: number | null;
  session_count_oos?: number;
  unique_imsi_count_no_coverage?: number;
  unique_imsi_count_no_coverage_with_duration_lte5min?: number;
  unique_imsi_count_no_coverage_with_duration_gt5min_lte15min?: number;
  unique_imsi_count_no_coverage_with_duration_gt15min_lte30min?: number;
  unique_imsi_count_no_coverage_with_duration_gt30min_lte1hour?: number;
  unique_imsi_count_no_coverage_with_duration_gt1hour_lte2hour?: number | null;
  unique_imsi_count_no_coverage_with_duration_gt2hour?: number | null;
  duration_no_coverage?: number;
  volume_no_coverage?: number;
  session_count_no_coverage?: number;
  unique_imsi_count_poor_coverage?: number;
  unique_imsi_count_poor_coverage_with_duration_lte5min?: number;
  unique_imsi_count_poor_coverage_with_duration_gt5min_lte15min?: number;
  unique_imsi_count_poor_coverage_with_duration_gt15min_lte30min?: number;
  unique_imsi_count_poor_coverage_with_duration_gt30min_lte1hour?: number;
  unique_imsi_count_poor_coverage_with_duration_gt1hour_lte2hour?: number;
  unique_imsi_count_poor_coverage_with_duration_gt2hour?: number | null;
  duration_poor_coverage?: number;
  volume_poor_coverage?: number;
  session_count_poor_coverage?: number;
  unique_imsi_count_medium_coverage?: number;
  unique_imsi_count_medium_coverage_with_duration_lte5min?: number;
  unique_imsi_count_medium_coverage_with_duration_gt5min_lte15min?: number;
  unique_imsi_count_medium_coverage_with_duration_gt15min_lte30min?: number;
  unique_imsi_count_medium_coverage_with_duration_gt30min_lte1hour?: number;
  unique_imsi_count_medium_coverage_with_duration_gt1hour_lte2hour?: number | null;
  unique_imsi_count_medium_coverage_with_duration_gt2hour?: number;
  duration_medium_coverage?: number;
  volume_medium_coverage?: number;
  session_count_medium_coverage?: number;
  unique_imsi_count_good_coverage?: number;
  unique_imsi_count_good_coverage_with_duration_lte5min?: number;
  unique_imsi_count_good_coverage_with_duration_gt5min_lte15min?: number;
  unique_imsi_count_good_coverage_with_duration_gt15min_lte30min?: number;
  unique_imsi_count_good_coverage_with_duration_gt30min_lte1hour?: number;
  unique_imsi_count_good_coverage_with_duration_gt1hour_lte2hour?: number;
  unique_imsi_count_good_coverage_with_duration_gt2hour?: number;
  duration_good_coverage?: number;
  volume_good_coverage?: number;
  session_count_good_coverage?: number;
  session_count_kddi?: number;
  unique_imsi_count_kddi?: number;
  polygon_used_for_analysis?: string;
  polygon_given?: string;
  units?: string;
  place_name?: string;
}

export interface CoverageResponse {
  query_type: string;
  query_input: string;
  received_at: string;
  coverage_items: CoverageItem[];
  raw_response: any[];
}